Added block & data generating code

+++ src/typing/types.rs
@@ -32,7 +32,7 @@ impl IValType for FloatTypeKind {
     fn gc_trace(&self, _: impl Fn(ValTypeID)) {}
 }
 impl FloatTypeKind {
-    pub fn get_binary_bits(&self) -> usize {
+    pub fn get_binary_bits(&self) -> u8 {
         match self {
             Self::Ieee32 => 32,
             Self::Ieee64 => 64,
